Updated game messages values
MSG_SELECT_BATTLECMD, MSG_SELECT_IDLECMD, MSG_SELECT_SUM, MSG_SELECT_CHAIN, MSG_SORT_CARD, MSG_SORT_CHAIN, MSG_CONFIRM_DECKTOP, MSG_CONFIRM_EXTRATOP, MSG_CONFIRM_CARDS,
MSG_SHUFFLE_HAND, MSG_SHUFFLE_EXTRA, MSG_CARD_SELECTED, MSG_RANDOM_SELECTED, MSG_BECOME_TARGET, MSG_DRAW:
Changed card counts values from int8 to int32.

MSG_SELECT_SUM, MSG_SELECT_CARD, MSG_SELECT_TRIBUTE, MSG_SELECT_UNSELECT_CARD:
Changed max and min value from int8 to int32

MSG_DECK_TOP, MSG_CHAINING:
Changed card sequence from int8 to int32.

MSG_TAG_SWAP:
Changed card counts values from int8 to int32.

MSG_RELOAD_FIELD:
Changed card counts values from int16 to int32.

+++ field.h
@@ -53,7 +53,7 @@ struct chain {
 	uint8 triggering_player;
 	uint8 triggering_controler;
 	uint16 triggering_location;
-	uint8 triggering_sequence;
+	uint32 triggering_sequence;
 	uint8 triggering_position;
 	card_state triggering_state;
 	effect* triggering_effect;
